# docs/03-development/api/openapi.yaml
openapi: 3.0.3
info:
  title: Novel Agent Studio API
  description: |
    多Agent协作小说创作平台API文档
    
    ## 功能特性
    - 多Agent协作章节生成
    - 世界观管理
    - 草稿持久化
    - Word导出
    - Agent Room实时对话
    
    ## 版本历史
    | 版本 | 日期 | 说明 |
    |------|------|------|
    | 1.0.0 | 2024-01-01 | 初始版本 |
  version: 1.0.0
  contact:
    name: Novel Agent Studio Team
servers:
  - url: http://localhost:8000
    description: 本地开发环境
  - url: https://api.novel-agent.com
    description: 生产环境
...
paths:
...
  /generate_chapter:
    post:
      tags:
        - generate
      summary: 完整章节生成
      description: |
        执行完整的Agent工作流生成章节：
        Planner -> Conflict -> Writing -> Editor -> Reader -> Summary
      operationId: generateChapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateChapterRequest'
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateChapterResponse'
        '422':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/novel/generate:
    post:
      tags:
        - generate
      summary: 简版章节生成
      description: 简化版生成接口，仅返回最终结果
      operationId: generateSimple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: 章节大纲
              required:
                - text
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  input:
                    type: string
                  result:
                    type: string
...
components:
  schemas:
...
    GenerateChapterRequest:
      type: object
      properties:
        outline:
          type: string
          description: 章节大纲
        story_id:
          type: string
          description: 故事ID
          default: demo-story
        agent_configs:
          type: object
          description: Agent配置参数
        constraints:
          type: array
          items:
            type: string
          description: 写作约束
        llm_config:
          type: object
          description: LLM配置
      required:
        - outline
    GenerateChapterResponse:
      type: object
      properties:
        input_text:
          type: string
        plan_text:
          type: string
        conflict_suggestions:
          type: array
          items:
            type: string
        draft_text:
          type: string
        edited_text:
          type: string
        reader_feedback:
          type: array
          items:
            type: string
        summary_text:
          type: string
        final_text:
          type: string
        agent_logs:
          type: array
          items:
            type: object
        trace_data:
          type: array
          items:
            $ref: '#/components/schemas/TraceText'
        story_id:
          type: string
...
    TraceText:
      type: object
      properties:
        text:
          type: string
        source_agent:
          type: string
        revisions:
          type: array
          items:
            type: string
...
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
              msg:
                type: string
              type:
                type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer